backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pymongo==4.9.0
motor==3.6.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
pytest==7.3.1
pytest-asyncio==0.21.0
aiohttp==3.9.1
